Added ProgramUnuse and BufferUnbind functions. Added *PointerTyped for explicitly specifing type of passed data. Added *PointerVBO for VBO. And one missing Uniform call.
(Color|Normal|...)PointerTyped is useful when one have interleaved buffer and it's stored in the array. Or for any other case where someone use byte array, but actual type of component is different.

(Color|Normal|...)PointerVBO is for case where gl*Pointer last argument is actually an offset inside VBO.