Implements numerous API changes.
These all involve functions which take slices and expect them
to be of a fixed size. These have been changed to accept a
pointer to a fixed-size array. This prevents possible memory
corruptions, while ensuring that type and capacity checks are
kept at compile-time, instead of run-time.

A number of functions which contained nothing but
'panic("Unimplemented")`, have now been implemented.

It should be noted that this patch will very likely break a lot
of code out there. But in the interest of getting the go-gl
toolset to a stable and useful state, we deemed this worth doing.